+++ openapi-diff/src/modeler/AutoRest.Swagger/Model/SwaggerObject.cs
@@ -82,6 +82,9 @@ namespace AutoRest.Swagger.Model
         [JsonProperty(PropertyName = "x-ms-enum")]
         public virtual XmsEnumExtension XmsEnum {get;set;}
 
+        [JsonProperty(PropertyName = "x-ms-client-name")]
+        public virtual string XmsClientName { get; set; }
+
         /// <summary>
         /// Compare a modified document node (this) to a previous one and look for breaking as well as non-breaking changes.
         /// </summary>
@@ -107,7 +110,9 @@ namespace AutoRest.Swagger.Model
 
             base.Compare(context, previous);
 
-            if (Reference != null && !Reference.Equals(prior.Reference))
+            var thisModelName = this.XmsClientName ?? Reference ?? "";
+            var priorModelName = previous.XmsClientName ?? previous.Reference ?? "";
+            if (!thisModelName.Equals(priorModelName))
             {
                 context.LogBreakingChange(ComparisonMessages.ReferenceRedirection);
             }